refactor: split user routes
