+++ docs/dev/ProjectBase.md
@@ -24,6 +24,7 @@
 - **Docker**：利用 Docker 打包 Node.js 服务、数据库以及其他依赖，实现一致的部署环境。
 - **HTTPS**：使用如 Let’s Encrypt 的 TLS 证书，保证通信安全。
 - **可扩展性**：API 设计采用模块化路由，便于未来新增服务。
+- **代码结构**：普通用户的登录、验证及凭据保存等接口已集中到 `server/users.js`，更易于维护。
 - **监控与日志**：使用 **Winston** 等日志工具及 **Prometheus** 之类的监控方案，跟踪系统状态。
 
 ## 后续步骤
